Candidate for 4.0.26

+++ src/main/scala/org/ergoplatform/mining/CandidateGenerator.scala
@@ -579,7 +579,7 @@ object CandidateGenerator extends ScorexLogging {
   /**
     * Transaction and its cost.
     */
-  type CostedTransaction = (ErgoTransaction, Long)
+  type CostedTransaction = (ErgoTransaction, Int)
 
   //TODO move ErgoMiner to mining package and make `collectTxs` and `fixTxsConflicts` private[mining]
 
@@ -680,8 +680,8 @@ object CandidateGenerator extends ScorexLogging {
     */
   def collectTxs(
     minerPk: ProveDlog,
-    maxBlockCost: Long,
-    maxBlockSize: Long,
+    maxBlockCost: Int,
+    maxBlockSize: Int,
     us: UtxoStateReader,
     upcomingContext: ErgoStateContext,
     transactions: Seq[ErgoTransaction]
